% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_coding.R
\name{ang_ang_plot}
\alias{ang_ang_plot}
\title{ang_ang_plot}
\usage{
ang_ang_plot(ang1, ang2)
}
\arguments{
\item{ang1}{Hip angle vector.}

\item{ang2}{Knee angle vector.}
}
\value{
A angle_angle figure.
}
\description{
angle-angle plot
}
\examples{

v1 <- sample$hip
v2 <- sample$knee
v3 <- sample$ankle
ang_ang_plot(v1, v2)
}
