% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclogram_area.R
\name{area}
\alias{area}
\title{Area of cyclogram}
\usage{
area(ang1, ang2, len)
}
\arguments{
\item{ang1}{Any body segment joint angle vector.}

\item{ang2}{Another body segment joint angle vector.}

\item{len}{Length of the input vector}
}
\value{
A numeric value of the area of cyclogram.
}
\description{
The cyclogram area is representative of the conjoint range of joint movements
}
\details{
area of cyclogram
}
\examples{
data(sample)
v1 <- sample$hip
v2 <- sample$knee
len <- length(v1)
area(v1, v2, len)
}
