% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_coding.R
\name{CAV}
\alias{CAV}
\title{CAV}
\usage{
CAV(df)
}
\arguments{
\item{df}{hip, ankle, knee dataframe}
}
\value{
A variability plot
}
\description{
CAV plot
}
\examples{
data(sample)
v1 <- sample$hip
v2 <- sample$knee
v3 <- sample$ankle
df <- sample
CAV(df)
ts.plot(CAV(df))
}
