% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclogram_area.R
\name{perimeter}
\alias{perimeter}
\title{perimeter of the cyclogram}
\usage{
perimeter(ang1, ang2, len)
}
\arguments{
\item{ang1}{Any body segment joint angle vector.}

\item{ang2}{Another body segment joint angle vector.}

\item{len}{Length of the input vector}
}
\value{
A numeric value of the perimeter of cyclogram.
}
\description{
The cyclogram perimeter provides information on the average joint velocity
}
\details{
perimeter of the cyclogram
}
\examples{
data(sample)
v1 <- sample$hip
v2 <- sample$knee
len <- length(v1)
perimeter(v1, v2, len)
}
