% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector coding 2.R
\name{phase_angle}
\alias{phase_angle}
\title{coupling angle plot}
\usage{
phase_angle(ang1, ang2)
}
\arguments{
\item{ang1}{Hip angle vector.}

\item{ang2}{Knee angle vector.}
}
\value{
A phase_angle figure.
}
\description{
coupling angle plot
}
\examples{
phase_angle(c(31.1977, 30.14246, 29.17131), c(22.40979, 24.59055, 26.00155))
}
