% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector coding 2.R
\name{phase_ratio}
\alias{phase_ratio}
\title{phase-ratio statistics}
\usage{
phase_ratio(pa)
}
\arguments{
\item{pa}{phase-angle vector.}
}
\value{
A phase-ratio dataframe.
}
\description{
phase-ratio statistics
}
\examples{
phase_ratio(c(11, 14, 158, 200, 544, 350, 47, 60, 240, 230, 270, 288, 300, 130))
}
