% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{data}
\name{sample}
\alias{sample}
\title{sample dataset}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{knee}{time series data recoding knee joint angles}
 \item{hip}{time series data recoding hip joint angles}
 \item{ankle}{time series data recoding ankle joint angles}
}
}
\usage{
data(sample)
}
\description{
sample dataset
}
\examples{

data(sample)
head(sample)
}
\keyword{datasets}
