% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector coding 2.R
\name{ang_ang_plot}
\alias{ang_ang_plot}
\title{angle-angle plot}
\usage{
ang_ang_plot(ang1, ang2)
}
\arguments{
\item{ang1}{Hip angle vector.}

\item{ang2}{Knee angle vector.}
}
\value{
A angle_angle figure.
}
\description{
angle-angle plot
}
\examples{
ang_ang_plot(c(31.1977, 30.14246, 29.17131), c(22.40979, 24.59055, 26.00155))
}
