% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_coding.R
\name{phase_angle}
\alias{phase_angle}
\title{phase_angle}
\usage{
phase_angle(ang1, ang2)
}
\arguments{
\item{ang1}{Any body segment joint angle vector.}

\item{ang2}{Another body segment joint angle vector.}
}
\value{
A phase_angle figure.
}
\description{
waited to read the paper
}
\details{
coupling angle plot
}
\examples{
data(sample)
v1 <- sample$hip
v2 <- sample$knee
v3 <- sample$ankle
phase_angle(v1, v2)
}
