% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{data}
\name{sample}
\alias{sample}
\title{sample dataset}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{knee}{}
 \item{hip}{}
 \item{ankle}{}
}
}
\usage{
data(sample)
}
\description{
sample dataset
}
\examples{

data(sample)
head(sample)
}
\keyword{datasets}
